Fixed Linux build error (must #include algorithm when using for_each).
While I was at it I removed a number of warnings (mostly unused variables).

+++ trunk/delta3d/src/dtCore/batchisector.cpp
@@ -239,14 +239,14 @@ namespace dtCore
    ///////////////////////////////////////////////////////////////////////////////
    void BatchIsector::SingleISector::GetHitPoint( osg::Vec3& xyz, int pointNum ) const
    {
-      if (pointNum >= GetNumberOfHits()) return;
+      if (pointNum >= (int) GetNumberOfHits()) return;
 
       xyz = mHitList[pointNum].getWorldIntersectPoint();
    }
    ///////////////////////////////////////////////////////////////////////////////
    void BatchIsector::SingleISector::GetHitPointNormal( osg::Vec3& normal, int pointNum ) const
    {
-      if (pointNum >= GetNumberOfHits()) return;
+      if (pointNum >= (int) GetNumberOfHits()) return;
 
       normal = mHitList[pointNum].getWorldIntersectNormal();
    }
